Logic: added code to convert the CSV into JSON The CSV is procured from the raw SQLite database from MySword.
The JSON is converted from the CSV by reading the whole file line by
line. Every Strong's number is converted into a JSON object key and the
lexical entry into its value. Along the way a couple of if functions are
added to cleanup the data and separate the Greek and Hebrew lexicon.

Other than that, the lexicon details are converted into a README file.

For more details check the comments in the script.